pyproject.toml
[project]
name = "2025-2026-holotrack"
version = "0.1.0"
description = "PFE holotrack"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "connected-components-3d>=3.10.0",
    "cupy-cuda12x>=13.0.0",
    "matplotlib==3.10.8",
    "napari==0.6.6",
    "numpy==2.4.2",
    "pandas>=2.2.0",
    "pillow==12.1.1",
    "pyyaml==6.0.3",
    "scipy==1.17.1",
    "setuptools==80.0.0",
    "tifffile>=2023.9.26",
    "torch==2.10.0",
    "torchmetrics>=1.7.0",
    "torchvision==0.25.0",
    "torchinfo>=1.8.0",
    "wandb>=0.24.0",
    "tqdm==4.67.3",
    "tensorboard>=2.20.0",
]

[tool.setuptools]
packages = ["holotrack_model", "visualizer", "Simulator"]
